Add ability to create both NetID and Email Users (#18)
# Release Notes

Adds the ability to create both NetID users _and_ external (email) users.

# Screenshot

<img width="313" alt="Screen Shot 2019-11-11 at 3 48 40 PM" src="https://user-images.githubusercontent.com/1099111/68595991-c718fe00-049a-11ea-988e-1984b0de038c.png">

NetID field becomes hidden when the checkbox is selected.

# Additional Context

I spent a good amount of time trying to unravel the `UsersController` to see if there was a good way to refactor it so that we could do this kind of thing up in open. Ultimately, I decided it wasn't worth the effort at this time.

The one thing I don't like here is that the form lives at `/facilities/:facility_id/users/new_external` even though you're not necessarily creating an external user. This is similar to what we do for Dartmouth, however, so I'm not that worried about it.

I would like to add a `username` format validation, but we'll need to know the format. That can, of course, be done separately.