UMass Speed Type Account Creation (#31)
# Release Notes

Add the ability to add a Speed Type Account to Corum. This first queries the API they provided us and only creates the account if it is valid.

# Additional Context

I lifted the majority of this from Dartmouth's `dc_chartstrings` engine and did some minor refactors. Theirs is a little more complicated because they have two account types while UMass only has one.

There's a separate story for a nightly job that fetches and updates the details, so I'll do that as a separate PR.