[#170385650] UMass Statement Tempate (#48)
# Release Notes

Initial formatting to match UMass's template for statements/invoices.

# Screenshot

Here is theirs:

![Screen Shot 2020-01-22 at 8 02 32 PM](https://user-images.githubusercontent.com/1099111/72950907-49068c80-3d52-11ea-8344-0815e3d40807.png)

Here is what it looks like after the changes in this PR:
<img width="810" alt="Screen Shot 2020-07-08 at 9 26 20 AM" src="https://user-images.githubusercontent.com/30355046/86931003-3a900f80-c0fd-11ea-81f4-e9db163d3653.png">


# Additional Context

I've never loved making these PDFs this way, but this is how we do it for the other schools. The code is kind of nasty because it's a lot of tables and styling. I took some liberties with not matching exact stylings from the example.

There are a couple todos still in code while I'm waiting to hear back from Charlene.

I also include the order number and date as columns which they don't have. I _think_, but I'm not sure they for Flow, they roll up everything into a line item or two in the invoice and then attach some logs that come from the machines. We will be changing their workflow a little bit.

There's a good chance this will be one of those things where we merge it in as is and make iterative tweaks as they get more familiar with the system.