[#155812] Check in eye.yml (#108)
# Release Notes

TECH TASK:
Check in eye.yml rather than rely on a template.
This makes it a little easier for maintainers to see the configuration and debug as needed.

For clarity we should delete `eye.yml` from the `shared` directory when this is merged in.

# Additional Context

The `auto_cancel` daemon is running on both prod servers right now.  I commented out the `auto_cancel` block on prod-01, when I do a prod release early next week we should be back to just running it on one server.