[#155284] Remove $0 journal rows from CSV and PRN exports (#128)
The system that UMass uses to upload journals into PeopleSoft does not handle rows for zero amounts. A single $0 row will block uploads for all other departments using the system (not just Corum). The current workaround has been to manually exclude the $0 rows by journaling them all together, and simply not submitting anything to PeopleSoft for those $0 journals. With the $0 transactions out of the way, a second journal is created and submitted to PeopleSoft.

This allows admins to generate CSV and PRN files for upload to PeopleSoft, even if they contain $0 order details.