[#156719] Allow ALS numbers to be reset annually (#129)
Added a table called `als_sequence_numbers` that should ONLY ever be used directly by the `AlsNumberGenerator`. When the max_value of 999 is reached, the Generator will throw an exception and the `als_sequence_numbers` will be reset by running the rake task: `rails umass_corum:reset_als_sequence`. We are using the id field of this new table as a faux db sequence since this version of MariaDB does not allow creating a standalone sequence. I believe this is the most thread-safe and consistent way of keeping track of the ALS sequence in the course of a year. Every July 1st we can run a cronjob that runs the `rails umass_corum:reset_als_sequence` rake task.

As part of this I also introduced a feature flag for the AlsNumberGenerator, which defaults to `true`. Until this is released, the ALS numbers will continue to be generated transiently via the `OrderDetailToJournalRowAttributes` class. Once released, the new als_number column on the `journals` table will start getting populated starting at "ALS001". We can then run the rake task `rails umass_corum:backfill_empty_als_numbers` rake task to save the ALS numbers generated before this change.