[#160820] Use ansible to allow the corum user to access ftp server (#381)
# Release Notes

I tried to use `community.general.ssh_config` for this but it required `stormssh` (and `pip`/`python3`) and I couldn't figure out how to specify a user that isn't actually on the server, which is the whole point of this config file.

In the end, I realized I could simply use `copy` to create the config file.  This works and it feels _pretty_ clear in its intent, even though it uses `copy` to create a file.