[#161290] Fix: implement dataprobe relay port configuration option (#387)
The TCP port field available on the product relay page is incorrectly hard-coded to 9100, which is the factory default port for this relay. This patch should be more useful, so that any valid user-defined integer will be used to report the status and to toggle the outlet for the product.