Updates `UmassCorum::OrderDetailStatementRowPresenter` to use `billable_duration_mins` (#562)
# Release Notes

Reservations for reconciled order details have their `billable_minutes` set to `nil`. UMass uses this attribute in their statements, which means the purchased quantity is blank. Dartmouth and NU use `PriceDisplayment#wrapped_quanity` to display the quantity on the statement. This method relies on `actual_duration_mins` and `duration_mins` instead of `billable_minutes`

UMass uses `UmassCorum::OrderDetailStatementRowPresenter#quantity` to display the quantity in invoices, and it depends on `billable_minutes`. This PR updates `quantity` to use `billable_duration_mins`, which uses a calculated value of, instead of the saved value for `billable_minutes`. Thus avoiding the `nil` value.